refactor(core): move Flare, context, and stacktrace into @flareapp/core
