+++ packages/core/src/context/collectAttributes.ts
