+++ packages/core/src/context/request.ts
@@ -1,11 +1,11 @@
 import { Attributes } from '../types';
-import { redactFullPath } from '../util';
+import { redactUrlQuery } from '../util';
 
 export default function request(urlDenylist: RegExp): Attributes {
     return {
-        'url.full': redactFullPath(window.location.href, urlDenylist),
+        'url.full': redactUrlQuery(window.location.href, urlDenylist),
         'user_agent.original': window.navigator.userAgent,
-        'http.request.referrer': redactFullPath(window.document.referrer, urlDenylist),
+        'http.request.referrer': redactUrlQuery(window.document.referrer, urlDenylist),
         'document.ready_state': window.document.readyState,
     };
 }
