+++ packages/core/src/stacktrace/createStackTrace.ts
