+++ packages/core/src/stacktrace/fileReader.ts
