+++ packages/core/src/stacktrace/index.ts
