+++ packages/core/src/stacktrace/nativeImport.ts
