+++ packages/js/src/index.ts
@@ -1,5 +1,6 @@
+import { Flare } from '@flareapp/core';
+
 import { catchWindowErrors } from './browser';
-import { Flare } from './Flare';
 
 export const flare = new Flare();
 
@@ -9,8 +10,9 @@ if (typeof window !== 'undefined' && window) {
     catchWindowErrors();
 }
 
-export { Flare } from './Flare';
-export { convertToError, DEFAULT_URL_DENYLIST, redactFullPath, resolveDenylist } from './util';
+export { Flare } from '@flareapp/core';
+export { convertToError, DEFAULT_URL_DENYLIST, redactUrlQuery, resolveDenylist } from '@flareapp/core';
+export { redactUrlQuery as redactFullPath } from '@flareapp/core';
 export type {
     AttributeValue,
     Attributes,
