+++ packages/js/tests/redactUrl.test.ts
@@ -1,8 +1,7 @@
+import { DEFAULT_URL_DENYLIST, Flare, redactUrlQuery as redactFullPath, resolveDenylist } from '@flareapp/core';
 // @vitest-environment jsdom
 import { afterEach, beforeEach, describe, expect, test } from 'vitest';
 
-import { Flare } from '../src/Flare';
-import { DEFAULT_URL_DENYLIST, redactFullPath, resolveDenylist } from '../src/util';
 import { FakeApi } from './helpers/FakeApi';
 
 describe('redactFullPath', () => {
